requirements.txt
django-tailwind == 3.8.0
django-tailwind[reload] == 3.8.0
channels == 4.2.0